Google Test 도입 및 단위 테스트 구축 + 빌드 에러 수정
- gtest 기반 Tests 프로젝트 추가 (Vector2, Actor, Level 테스트 22개)
- Actor.cpp GetPositionI() narrowing conversion 수정
- Engine.h Input forward declaration → #include 로 교체
- Level.h 복사 생성자/대입 연산자 = delete 추가
- Engine.cpp settings.frameRate stoi → static_cast<float> 수정
- Level::ProcessAddAndDestroyActor() O(n²) → swap-and-pop O(n) 개선

+++ Engine/Engine/Engine.cpp
@@ -123,7 +123,7 @@ namespace Engine
 				{
 					if (line.find("FrameRate:") == 0)
 					{
-						settings.frameRate = std::stoi(line.substr(10));
+						settings.frameRate = static_cast<float>(std::stoi(line.substr(10)));
 					}
 					else if (line.find("Width:") == 0)
 					{
